Makes reference to Back To The Future
Looking through some old web projects, I came across my favorite
learning tool. PEZ!

I hate to say it, but RSS is dying. Like, really dying. Some of the
stuff in Pez (Zep here because this is where I iron out all of the
wrinkles before I submit it to Pez over at Google Code. Something I
haven't done in a while!) no longer work like the ability to have your
tweets on your profile because Twitter killed off the RSS feed. Because
RSS is dead you know?

So where do I go from here? This is a project that was started in 2008,
and here we are, just over five years later, and I'm still slowly
tinkering away. Well, I have this idea for another project that I want
to work on, and this is where Pez comes in. I would like to look at
other ways of bringing in your tweets, Facebook status messages, and
even the photos of your food that you post to Instagram. I'll use Pez to
figure that out, and push that into the other project that I am working
on while still working on this zombie of a project that I had high hopes
for and I really feel it had a place in 2008, and it has a place now,
but it just needs an update.

Talking to myself,
Ben